refactor: filter 컴포넌트 수정 및 z-index 적용
