+++ src/components/Category&Filter/Filter/Filter.module.scss
@@ -1,4 +1,5 @@
 .container {
+  position: relative;
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -27,10 +28,14 @@
   }
 }
 .dropdown {
+  position: absolute;
+  top: 6.4rem;
+  z-index: 2;
   width: 100%;
   box-shadow: 0rem 0.4rem 1.6rem 0rem rgba(17, 34, 17, 0.05);
   border-radius: 0.6rem;
   border: 0.1rem solid $gray-300;
+  background-color: $white;
   .item {
     display: flex;
     justify-content: center;
