+++ src/components/Category&Filter/Filter/FilterList.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef } from 'react';
+import React from 'react';
 import classNames from 'classnames/bind';
 
 import styles from './Filter.module.scss';
@@ -11,27 +11,13 @@ interface filterListProps {
   setSelectedItem: React.Dispatch<React.SetStateAction<string>>;
 }
 export default function FilterList({ filterList, setIsOpen, setSelectedItem }: filterListProps) {
-  const filterRef = useRef<HTMLDivElement>(null);
-
   const handleItemClick = (index: number) => {
     setSelectedItem(filterList[index].name);
     setIsOpen(false);
   };
 
-  useEffect(() => {
-    const handleOutsideClick = (event: MouseEvent) => {
-      if (filterRef.current && !filterRef.current.contains(event.target as Node)) {
-        setIsOpen(false);
-      }
-    };
-    document.addEventListener('mousedown', handleOutsideClick);
-    return () => {
-      document.removeEventListener('mousedown', handleOutsideClick);
-    };
-  }, [setIsOpen]);
-
   return (
-    <div ref={filterRef} className={cn('dropdown')}>
+    <div className={cn('dropdown')}>
       {filterList.map((item, index) => (
         <div key={`${index} ${item.element}`} className={cn('item')} onClick={() => handleItemClick(index)}>
           {item.element}
